# docs/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "PDF Exporter REST API",
    "version": "v1"
  },
  "paths": {
...
    "/api/export-filename": {
      "post": {
        "operationId": "getFileName",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExportParams"
              }
            }
          },
          "description": "Export parameters similar as for generation of PDF",
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successfully generated filename"
          }
        },
        "summary": "Gets a filename, prepared with velocity and placeholders",
        "tags": [
          "Utility resources"
        ]
      }
    },
...
    "/api/projects/{projectId}/testruns/{testRunId}/attachments": {
      "get": {
        "operationId": "getTestRunAttachments",
        "parameters": [
          {
            "description": "Project ID",
            "in": "path",
            "name": "projectId",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "TestRun ID",
            "in": "path",
            "name": "testRunId",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "TestRun revision",
            "in": "query",
            "name": "revision",
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "Filename filter for attachment",
            "in": "query",
            "name": "filter",
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "Testcase boolean filter field ID",
            "in": "query",
            "name": "testCaseFilterFieldId",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/TestRunAttachment"
                  },
                  "type": "array"
                }
              }
            },
            "description": "Successfully retrieved list of attachments"
          }
        },
        "summary": "Gets list of attachments of specified workitem using optional filter",
        "tags": [
          "TestRuns"
        ]
      }
    },
    "/api/projects/{projectId}/testruns/{testRunId}/attachments/{attachmentId}": {
      "get": {
        "operationId": "getTesRunAttachment",
        "parameters": [
          {
            "description": "Project ID",
            "in": "path",
            "name": "projectId",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "TestRun ID",
            "in": "path",
            "name": "testRunId",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "Attachment ID",
            "in": "path",
            "name": "attachmentId",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "Attachment revision",
            "in": "query",
            "name": "revision",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TestRunAttachment"
                }
              }
            },
            "description": "Successfully retrieved attachment"
          }
        },
        "summary": "Gets attachment of specified workitem",
        "tags": [
          "TestRuns"
        ]
      }
    },
    "/api/projects/{projectId}/testruns/{testRunId}/attachments/{attachmentId}/content": {
      "get": {
        "operationId": "getTestRunAttachmentContent",
        "parameters": [
          {
            "description": "Project ID",
            "in": "path",
            "name": "projectId",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "TestRun ID",
            "in": "path",
            "name": "testRunId",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "Attachment ID",
            "in": "path",
            "name": "attachmentId",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "Attachment revision",
            "in": "query",
            "name": "revision",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully retrieved attachment content"
          }
        },
        "summary": "Gets content of attachment of specified workitem",
        "tags": [
          "TestRuns"
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ExportParams": {
        "properties": {
          "attachmentsFilter": {
            "description": "Filter for attachments to be downloaded, example: '*.pdf'",
            "type": "string"
          },
          "autoSelectStylePackage": {
            "description": "Indicator that most parameters from this object should be ignored, best suitable style package should be found and properties from it should be taken instead",
            "type": "boolean"
          },
          "baselineRevision": {
            "description": "The specific revision of the provided baseline",
            "type": "string"
          },
          "chapters": {
            "description": "Specific higher level chapters",
            "items": {
              "description": "Specific higher level chapters",
              "type": "string"
            },
            "type": "array"
          },
          "coverPage": {
            "description": "Cover page settings name",
            "type": "string"
          },
          "css": {
            "description": "CSS settings name",
            "type": "string"
          },
          "cutEmptyChapters": {
            "description": "Empty chapters should be removed from the document",
            "type": "boolean"
          },
          "cutEmptyWIAttributes": {
            "default": true,
            "description": "Empty work item attributes should be removed from the document",
            "type": "boolean"
          },
          "cutLocalUrls": {
            "description": "Local Polarion URLs should be removed from the document",
            "type": "boolean"
          },
          "documentType": {
            "description": "Type of the document",
            "enum": [
              "BASELINE_COLLECTION",
              "LIVE_DOC",
              "LIVE_REPORT",
              "TEST_RUN",
              "WIKI_PAGE"
            ],
            "example": "LIVE_DOC",
            "type": "string"
          },
          "embedAttachments": {
            "description": "If attachments should be embedded into PDF",
            "type": "boolean"
          },
          "fileName": {
            "description": "Target file name of exported document",
            "type": "string"
          },
          "fitToPage": {
            "default": true,
            "description": "Content should be scaled to fit the page",
            "type": "boolean"
          },
          "followHTMLPresentationalHints": {
            "default": true,
            "description": "HTML presentational hints should be followed",
            "type": "boolean"
          },
          "fullFonts": {
            "default": false,
            "description": "Embed full fonts instead of subsetting. When true, fonts are embedded completely without optimization. This can help avoid font subsetting errors but results in larger PDF files.",
            "type": "boolean"
          },
          "headerFooter": {
            "description": "Header/Footer settings name",
            "type": "string"
          },
          "headersColor": {
            "description": "Color to be used for headers in the document. By default dark blue color (Polarion's default)",
            "type": "string"
          },
          "imageDensity": {
            "description": "Quality of PNG images converted from SVG",
            "enum": [
              "DPI_96",
              "DPI_192",
              "DPI_300",
              "DPI_600"
            ],
            "type": "string"
          },
          "includeUnreferencedComments": {
            "description": "Include unreferenced comments",
            "type": "boolean"
          },
          "internalContent": {
            "description": "Internal content",
            "type": "string"
          },
          "language": {
            "description": "Language in the exported document",
            "type": "string"
          },
          "linkRoleDirection": {
            "description": "Direction of linked workitem roles resolution",
            "enum": [
              "BOTH",
              "DIRECT",
              "REVERSE"
            ],
            "type": "string"
          },
          "linkedWorkitemRoles": {
            "description": "Specific Workitem roles",
            "example": [
              "has parent",
              "depends on"
            ],
            "items": {
              "description": "Specific Workitem roles",
              "example": "[\"has parent\",\"depends on\"]",
              "type": "string"
            },
            "type": "array"
          },
          "localization": {
            "description": "Localization settings name",
            "type": "string"
          },
          "locationPath": {
            "description": "Document path for export",
            "example": "Specification/Product Specification",
            "type": "string"
          },
          "markReferencedWorkitems": {
            "description": "Referenced work items should be marked in the document",
            "type": "boolean"
          },
          "metadataFields": {
            "description": "CSV list of document field names or wildcard patterns to be exported as HTML meta (LiveDoc only)",
            "items": {
              "description": "CSV list of document field names or wildcard patterns to be exported as HTML meta (LiveDoc only)",
              "type": "string"
            },
            "type": "array"
          },
          "numberedListStyles": {
            "description": "SCC styles to be applied to numbered lists in the document",
            "type": "string"
          },
          "orientation": {
            "default": "PORTRAIT",
            "description": "Orientation of the output document",
            "enum": [
              "PORTRAIT",
              "LANDSCAPE"
            ],
            "type": "string"
          },
          "paperSize": {
            "default": "A4",
            "description": "Standard paper sizes",
            "enum": [
              "A5",
              "A4",
              "A3",
              "B5",
              "B4",
              "JIS_B5",
              "JIS_B4",
              "LETTER",
              "LEGAL",
              "LEDGER"
            ],
            "type": "string"
          },
          "pdfVariant": {
            "default": "PDF_A_2B",
            "description": "PDF variants",
            "enum": [
              "PDF_A_1A",
              "PDF_A_1B",
              "PDF_A_2A",
              "PDF_A_2B",
              "PDF_A_2U",
              "PDF_A_3A",
              "PDF_A_3B",
              "PDF_A_3U",
              "PDF_A_4E",
              "PDF_A_4F",
              "PDF_A_4U",
              "PDF_UA_1",
              "PDF_UA_2"
            ],
            "type": "string"
          },
          "projectId": {
            "description": "The unique identifier for the project",
            "example": "elibrary",
            "type": "string"
          },
          "renderComments": {
            "description": "Which comments should be rendered in the exported document",
            "enum": [
              "OPEN",
              "ALL"
            ],
            "type": "string"
          },
          "renderNativeComments": {
            "description": "Render native comments",
            "type": "boolean"
          },
          "revision": {
            "description": "The specific revision of the document",
            "type": "string"
          },
          "testcaseFieldId": {
            "description": "A boolean testcase field ID. Attachments will be downloaded only from the testcases which have True value in the provided field. Leaving field empty will process all testcases.",
            "type": "string"
          },
          "urlQueryParameters": {
            "additionalProperties": {
              "description": "Map of attributes extracted from the URL",
              "type": "string"
            },
            "description": "Map of attributes extracted from the URL",
            "type": "object"
          },
          "watermark": {
            "description": "Watermark content to be applied to the document",
            "type": "boolean"
          },
          "webhooks": {
            "description": "Webhooks settings name",
            "type": "string"
          }
        },
        "type": "object"
      },
...
      "TestRunAttachment": {
        "properties": {
          "fileName": {
            "type": "string"
          },
          "id": {
            "type": "string"
          },
          "revision": {
            "type": "string"
          }
        },
        "type": "object"
      },
...
    }
  },
...
}