# src/main/java/ch/sbb/polarion/extension/pdf_exporter/rest/controller/TestRunAttachmentsInternalController.java
package ch.sbb.polarion.extension.pdf_exporter.rest.controller;

import ch.sbb.polarion.extension.pdf_exporter.rest.model.attachments.TestRunAttachment;
import ch.sbb.polarion.extension.pdf_exporter.service.PdfExporterPolarionService;
import com.polarion.alm.tracker.model.ITestRunAttachment;
import io.swagger.v3.oas.annotations.Hidden;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.media.ArraySchema;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.media.Schema;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.tags.Tag;

import jakarta.inject.Singleton;
import jakarta.ws.rs.GET;
import jakarta.ws.rs.Path;
import jakarta.ws.rs.PathParam;
import jakarta.ws.rs.Produces;
import jakarta.ws.rs.QueryParam;
import jakarta.ws.rs.core.HttpHeaders;
import jakarta.ws.rs.core.MediaType;
import jakarta.ws.rs.core.Response;
import java.util.List;

@Singleton
@Hidden
@Path("/internal")
@Tag(name = "TestRuns")
public class TestRunAttachmentsInternalController {

    private static final String FILENAME_HEADER = "Filename";

    private final PdfExporterPolarionService pdfExporterPolarionService;

    public TestRunAttachmentsInternalController() {
        pdfExporterPolarionService = new PdfExporterPolarionService();
    }

    public TestRunAttachmentsInternalController(PdfExporterPolarionService pdfExporterPolarionService) {
        this.pdfExporterPolarionService = pdfExporterPolarionService;
    }

    @GET
    @Path("/projects/{projectId}/testruns/{testRunId}/attachments")
    @Produces(MediaType.APPLICATION_JSON)
    @Operation(summary = "Gets list of attachments of specified workitem using optional filter",
            responses = {
                    @ApiResponse(responseCode = "200",
                            description = "Successfully retrieved list of attachments",
                            content = @Content(array = @ArraySchema(schema = @Schema(implementation = TestRunAttachment.class)))
                    )
            }
    )
    public List<TestRunAttachment> getTestRunAttachments(
            @Parameter(description = "Project ID", required = true) @PathParam("projectId") String projectId,
            @Parameter(description = "TestRun ID", required = true) @PathParam("testRunId") String testRunId,
            @Parameter(description = "TestRun revision") @QueryParam("revision") String revision,
            @Parameter(description = "Filename filter for attachment") @QueryParam("filter") String filter,
            @Parameter(description = "Testcase boolean filter field ID") @QueryParam("testCaseFilterFieldId") String testCaseFilterFieldId
    ) {
        return pdfExporterPolarionService.getTestRunAttachments(projectId, testRunId, revision, filter, testCaseFilterFieldId);
    }

    @GET
    @Path("/projects/{projectId}/testruns/{testRunId}/attachments/{attachmentId}")
    @Produces(MediaType.APPLICATION_JSON)
    @Operation(summary = "Gets attachment of specified workitem",
            responses = {
                    @ApiResponse(responseCode = "200",
                            description = "Successfully retrieved attachment",
                            content = @Content(schema = @Schema(implementation = TestRunAttachment.class))
                    )
            }
    )
    public TestRunAttachment getTesRunAttachment(
            @Parameter(description = "Project ID", required = true) @PathParam("projectId") String projectId,
            @Parameter(description = "TestRun ID", required = true) @PathParam("testRunId") String testRunId,
            @Parameter(description = "Attachment ID", required = true) @PathParam("attachmentId") String attachmentId,
            @Parameter(description = "Attachment revision") @QueryParam("revision") String revision
    ) {
        ITestRunAttachment testRunAttachment = pdfExporterPolarionService.getTestRunAttachment(projectId, testRunId, attachmentId, revision);
        return TestRunAttachment.fromAttachment(testRunAttachment);
    }

    @GET
    @Path("/projects/{projectId}/testruns/{testRunId}/attachments/{attachmentId}/content")
    @Produces(MediaType.APPLICATION_OCTET_STREAM)
    @Operation(summary = "Gets content of attachment of specified workitem",
            responses = {
                    @ApiResponse(responseCode = "200", description = "Successfully retrieved attachment content")
            }
    )
    public Response getTestRunAttachmentContent(
            @Parameter(description = "Project ID", required = true) @PathParam("projectId") String projectId,
            @Parameter(description = "TestRun ID", required = true) @PathParam("testRunId") String testRunId,
            @Parameter(description = "Attachment ID", required = true) @PathParam("attachmentId") String attachmentId,
            @Parameter(description = "Attachment revision") @QueryParam("revision") String revision
    ) {
        ITestRunAttachment testRunAttachment = pdfExporterPolarionService.getTestRunAttachment(projectId, testRunId, attachmentId, revision);

        return Response.ok(testRunAttachment.getDataStream())
                .header(HttpHeaders.CONTENT_DISPOSITION, "attachment; filename=\"" + testRunAttachment.getFileName() + "\"")
                .header(FILENAME_HEADER, testRunAttachment.getFileName())
                .build();
    }
}